Remove old wrappers around ForwardDiff, ReverseDiff and Mooncake. Use DifferentiationInterface instead
